+++ Project.toml
@@ -41,12 +41,12 @@ Configurations = "0"
 CountFlops = "0"
 DataFrames = "1"
 HTTP = "1"
-Hwloc = "3.3.0"
+Hwloc = "3.3"
 JLD2 = "0"
 JSON = "0"
 MLStyle = "0"
 MacroTools = "0"
-PrecompileTools = "1.2.1"
+PrecompileTools = "1.2"
 Suppressor = "0"
-ThreadPinningCore = "0.4.5"
+ThreadPinningCore = "0.4"
 UnicodePlots = "3"
