+++ examples/example-paper-implicitglobalgrid/perftest_config.toml
@@ -6,22 +6,24 @@ mode = "reduce"
 enabled = false
 
 [general]
-recursive = true
 save_results = true
-autoflops = false
-suppress_output = true
-plotting = true
-save_folder = ".perftests"
 max_saved_results = 20
-verbose = true
+suppress_output = true
+save_folder = ".perftests"
+threads_per_numa = "single"
+recursive = true
+verbose = 0
+autoflops = true
+plotting = true
+numas = "single"
 logs_enabled = true
 safe_formulas = false
 
 [regression]
-enabled = false
-default_threshold = 0.9
-use_bencher = false
+enabled = true
 dedicated_reference_file = ""
+default_threshold = 1.1
+use_bencher = false
 
 [roofline]
 enabled = true
