+++ examples/example-quickstart/Project.toml
@@ -1,0 +1,2 @@
+[deps]
+PerfTest = "1dca261b-fc56-4a8c-a7e2-9798d8a75978"
