+++ src/transform/configuration.jl
@@ -197,7 +197,7 @@ CONFIG_SHAPE = Dict(
 
 DEFAULT = Dict(
     "general" => Dict(
-        "autoflops" => true,
+        "autoflops" => false,
         "numas" => "single",
         "threads_per_numa" => "single",
         "save_results" => true,
@@ -246,8 +246,8 @@ DEFAULT = Dict(
 PRECOMPILATION_CONFIG = Dict(
     "general" => Dict(
         "autoflops" => false,
-        "numas" => "single",
-        "threads_per_numa" => "single",
+        "numas" => 1,
+        "threads_per_numa" => 1,
         "save_results" => false,
         "logs_enabled" => false,
         "save_folder" => ".thisshouldnotexist",
